CM-1597: Extracted asynchronous log methods into BaseExperimentAsync to clearly separate asynchronous logic from synchronous. Fixed related test cases
